+++ sdk/mysql/azure-resourcemanager-mysql/src/main/java/com/azure/resourcemanager/mysql/implementation/DatabasesClientImpl.java
@@ -84,7 +84,7 @@ public final class DatabasesClientImpl implements DatabasesClient {
             @HeaderParam("Accept") String accept,
             Context context);
 
-        @Headers({"Accept: application/json;q=0.9", "Content-Type: application/json"})
+        @Headers({"Content-Type: application/json"})
         @Delete(
             "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.DBForMySQL/servers"
                 + "/{serverName}/databases/{databaseName}")
@@ -97,6 +97,7 @@ public final class DatabasesClientImpl implements DatabasesClient {
             @PathParam("resourceGroupName") String resourceGroupName,
             @PathParam("serverName") String serverName,
             @PathParam("databaseName") String databaseName,
+            @HeaderParam("Accept") String accept,
             Context context);
 
         @Headers({"Content-Type: application/json"})
@@ -452,6 +453,7 @@ public final class DatabasesClientImpl implements DatabasesClient {
             return Mono.error(new IllegalArgumentException("Parameter databaseName is required and cannot be null."));
         }
         final String apiVersion = "2017-12-01";
+        final String accept = "application/json";
         return FluxUtil
             .withContext(
                 context ->
@@ -463,6 +465,7 @@ public final class DatabasesClientImpl implements DatabasesClient {
                             resourceGroupName,
                             serverName,
                             databaseName,
+                            accept,
                             context))
             .subscriberContext(context -> context.putAll(FluxUtil.toReactorContext(this.client.getContext())));
     }
@@ -505,6 +508,7 @@ public final class DatabasesClientImpl implements DatabasesClient {
             return Mono.error(new IllegalArgumentException("Parameter databaseName is required and cannot be null."));
         }
         final String apiVersion = "2017-12-01";
+        final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service
             .delete(
@@ -514,6 +518,7 @@ public final class DatabasesClientImpl implements DatabasesClient {
                 resourceGroupName,
                 serverName,
                 databaseName,
+                accept,
                 context);
     }