+++ sdk/mysql/azure-resourcemanager-mysql/src/main/java/com/azure/resourcemanager/mysql/implementation/FirewallRulesClientImpl.java
@@ -84,7 +84,7 @@ public final class FirewallRulesClientImpl implements FirewallRulesClient {
             @HeaderParam("Accept") String accept,
             Context context);
 
-        @Headers({"Accept: application/json;q=0.9", "Content-Type: application/json"})
+        @Headers({"Content-Type: application/json"})
         @Delete(
             "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.DBForMySQL/servers"
                 + "/{serverName}/firewallRules/{firewallRuleName}")
@@ -97,6 +97,7 @@ public final class FirewallRulesClientImpl implements FirewallRulesClient {
             @PathParam("resourceGroupName") String resourceGroupName,
             @PathParam("serverName") String serverName,
             @PathParam("firewallRuleName") String firewallRuleName,
+            @HeaderParam("Accept") String accept,
             Context context);
 
         @Headers({"Content-Type: application/json"})
@@ -475,6 +476,7 @@ public final class FirewallRulesClientImpl implements FirewallRulesClient {
                 .error(new IllegalArgumentException("Parameter firewallRuleName is required and cannot be null."));
         }
         final String apiVersion = "2017-12-01";
+        final String accept = "application/json";
         return FluxUtil
             .withContext(
                 context ->
@@ -486,6 +488,7 @@ public final class FirewallRulesClientImpl implements FirewallRulesClient {
                             resourceGroupName,
                             serverName,
                             firewallRuleName,
+                            accept,
                             context))
             .subscriberContext(context -> context.putAll(FluxUtil.toReactorContext(this.client.getContext())));
     }
@@ -529,6 +532,7 @@ public final class FirewallRulesClientImpl implements FirewallRulesClient {
                 .error(new IllegalArgumentException("Parameter firewallRuleName is required and cannot be null."));
         }
         final String apiVersion = "2017-12-01";
+        final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service
             .delete(
@@ -538,6 +542,7 @@ public final class FirewallRulesClientImpl implements FirewallRulesClient {
                 resourceGroupName,
                 serverName,
                 firewallRuleName,
+                accept,
                 context);
     }