+++ sdk/mysql/azure-resourcemanager-mysql/src/main/java/com/azure/resourcemanager/mysql/implementation/ServerAdministratorsClientImpl.java
@@ -71,7 +71,7 @@ public final class ServerAdministratorsClientImpl implements ServerAdministrator
         @Headers({"Content-Type: application/json"})
         @Get(
             "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.DBForMySQL/servers"
-                + "/{serverName}/Administrators/activeDirectory")
+                + "/{serverName}/administrators/activeDirectory")
         @ExpectedResponses({200})
         @UnexpectedResponseExceptionType(ManagementException.class)
         Mono<Response<ServerAdministratorResourceInner>> get(
@@ -86,7 +86,7 @@ public final class ServerAdministratorsClientImpl implements ServerAdministrator
         @Headers({"Content-Type: application/json"})
         @Put(
             "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.DBForMySQL/servers"
-                + "/{serverName}/Administrators/activeDirectory")
+                + "/{serverName}/administrators/activeDirectory")
         @ExpectedResponses({200, 202})
         @UnexpectedResponseExceptionType(ManagementException.class)
         Mono<Response<Flux<ByteBuffer>>> createOrUpdate(
@@ -102,7 +102,7 @@ public final class ServerAdministratorsClientImpl implements ServerAdministrator
         @Headers({"Content-Type: application/json"})
         @Delete(
             "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.DBForMySQL/servers"
-                + "/{serverName}/Administrators/activeDirectory")
+                + "/{serverName}/administrators/activeDirectory")
         @ExpectedResponses({200, 202, 204})
         @UnexpectedResponseExceptionType(ManagementException.class)
         Mono<Response<Flux<ByteBuffer>>> delete(